+++ src/system/documents/system-embedded-collections/constants.ts
@@ -1,0 +1,2 @@
+export const SYSTEM_EMBEDDED_COLLECTIONS_KEY =
+    '__systemEmbeddedCollections' as const;
